requirements.txt
# Main application framework
streamlit>=1.28.0

# Face recognition and image handling
face-recognition>=1.3.0
dlib>=19.24.0
numpy>=1.24.0
Pillow>=10.0.0

# Database
SQLAlchemy>=2.0.0

# Encryption
cryptography>=41.0.0

# Project tracking
watchdog>=4.0.0

# Voice
gTTS>=2.5.0
SpeechRecognition>=3.10.0
PyAudio>=0.2.14
openai-whisper>=20231117

# Local LLM
requests>=2.31.0

# Local Image Gen
diffusers>=0.27.0
transformers>=4.40.0
torch>=2.3.0